+++ R/extract_draws.R
@@ -34,10 +34,10 @@ extract_draws.brmsfit <- function(x, newdata = NULL, re_formula = NULL,
   meef <- tidy_meef(bterms, x$data)
   old_sdata <- NULL
   if (new) {
-    # extract_draws_re() also requires the new level names
+    # extract_draws_re() also requires the levels from newdata
     # original level names are already passed via old_ranef
-    new_levels <- attr(tidy_ranef(bterms, newdata), "levels")
-    attr(ranef, "levels") <- new_levels
+    used_levels <- attr(tidy_ranef(bterms, newdata), "levels")
+    attr(ranef, "levels") <- used_levels
     if (length(get_effect(bterms, "gp"))) {
       # GPs for new data require the original data as well
       old_sdata <- standata(x, internal = TRUE, ...)
@@ -493,7 +493,8 @@ extract_draws_re <- function(bterms, samples, sdata, data, ranef, old_ranef,
                              sample_new_levels = "uncertainty", ...) {
   # extract draws of group-level effects
   # Args:,
-  #   ranef: output of tidy_ranef based on the new formula and old data
+  #   ranef: output of 'tidy_ranef' based on the new formula and old data
+  #          but storing levels obtained from new data
   #   old_ranef: same as 'ranef' but based on the original formula
   draws <- list()
   px <- check_prefix(bterms)
@@ -501,175 +502,93 @@ extract_draws_re <- function(bterms, samples, sdata, data, ranef, old_ranef,
   if (!nrow(ranef)) {
     return(draws)
   }
-  p <- combine_prefix(px)
+  p <- usc(combine_prefix(px))
   groups <- unique(ranef$group)
   old_ranef <- subset2(old_ranef, ls = px)
   # assigning S4 objects requires initialisation of list elements
   draws[c("Z", "Zsp", "Zcs")] <- list(named_list(groups))
   for (g in groups) {
     # prepare general variables related to group g
-    sub_ranef <- subset2(ranef, group = g)
-    sub_old_ranef <- subset2(old_ranef, group = g)
-    new_levels <- attr(ranef, "levels")[[g]]
+    g_ranef <- subset2(ranef, group = g)
+    g_old_ranef <- subset2(old_ranef, group = g)
+    used_levels <- attr(ranef, "levels")[[g]]
     old_levels <- attr(old_ranef, "levels")[[g]]
-    new_by_per_level <- attr(new_levels, "by")
-    old_by_per_level <- attr(old_levels, "by")
-    really_new_levels <- setdiff(new_levels, old_levels)
     nlevels <- length(old_levels) 
-    nranef <- nrow(sub_ranef)
+    nranef <- nrow(g_ranef)
     # prepare samples of group-level effects
-    rpars <- paste0("^r_", g, usc(usc(p)), "\\[")
+    rpars <- paste0("^r_", g, usc(p), "\\[")
     rsamples <- get_samples(samples, rpars)
-    if (is.null(rsamples)) {
+    if (!length(rsamples)) {
       stop2(
         "Group-level effects for each level of group ", 
         "'", g, "' not found. Please set 'save_ranef' to ", 
         "TRUE when fitting your model."
       )
     }
-    new_rpars <- match(sub_ranef$coef, sub_old_ranef$coef)
-    new_rpars <- outer(seq_len(nlevels), (new_rpars - 1) * nlevels, "+")
-    new_rpars <- as.vector(new_rpars)
-    rsamples <- rsamples[, new_rpars, drop = FALSE]
+    # only use samples of effects specified in the new formula
+    used_rpars <- match(g_ranef$coef, g_old_ranef$coef)
+    used_rpars <- outer(seq_len(nlevels), (used_rpars - 1) * nlevels, "+")
+    used_rpars <- as.vector(used_rpars)
+    rsamples <- rsamples[, used_rpars, drop = FALSE]
+    rsamples <- column_to_row_major_order(rsamples, nranef)
     # prepare data required for indexing parameters
-    gtype <- sub_ranef$gtype[1]
-    id <- sub_ranef$id[1]
-    idresp <- paste0(id, usc(sub_ranef$resp[1]))
+    gtype <- g_ranef$gtype[1]
+    id <- g_ranef$id[1]
+    idresp <- paste0(id, usc(g_ranef$resp[1]))
     if (gtype == "mm") {
-      ngf <- length(sub_ranef$gcall[[1]]$groups)
+      ngf <- length(g_ranef$gcall[[1]]$groups)
       gf <- sdata[paste0("J_", idresp, "_", seq_len(ngf))]
       weights <- sdata[paste0("W_", idresp, "_", seq_len(ngf))]
     } else {
       gf <- sdata[paste0("J_", idresp)]
       weights <- list(rep(1, length(gf[[1]])))
     }
-    # incorporate new gf levels
-    new_rsamples <- vector("list", length(gf))
-    max_level <- nlevels
-    for (i in seq_along(gf)) {
-      has_new_levels <- any(gf[[i]] > nlevels)
-      if (has_new_levels) {
-        if (sample_new_levels %in% c("old_levels", "gaussian")) {
-          new_indices <- sort(setdiff(gf[[i]], seq_len(nlevels)))
-          new_rsamples[[i]] <- matrix(
-            nrow = nrow(rsamples), ncol = nranef * length(new_indices)
-          )
-          if (sample_new_levels == "old_levels") {
-            for (j in seq_along(new_indices)) {
-              # choose a person to take the group-level effects from
-              if (length(old_by_per_level)) {
-                new_by <- new_by_per_level[new_levels == really_new_levels[j]]
-                possible_levels <- old_levels[old_by_per_level == new_by]
-                possible_levels <- which(old_levels %in% possible_levels)
-                take_level <- sample(possible_levels, 1)
-              } else {
-                take_level <- sample(seq_len(nlevels), 1)
-              }
-              for (k in seq_len(nranef)) {
-                take <- (k - 1) * nlevels + take_level
-                new_rsamples[[i]][, (j - 1) * nranef + k] <- rsamples[, take]
-              }
-            }
-          } else if (sample_new_levels == "gaussian") {
-            if (any(!sub_ranef$dist %in% "gaussian")) {
-              stop2("Option sample_new_levels = 'gaussian' is not ",
-                    "available for non-gaussian group-level effects.")
-            }
-            for (j in seq_along(new_indices)) {
-              # extract hyperparameters used to compute the covariance matrix
-              if (length(old_by_per_level)) {
-                new_by <- new_by_per_level[new_levels == really_new_levels[j]]
-                rnames <- as.vector(get_rnames(sub_ranef, bylevels = new_by))
-              } else {
-                rnames <- get_rnames(sub_ranef)
-              }
-              sd_pars <- paste0("sd_", g, "__", rnames)
-              sd_samples <- get_samples(samples, sd_pars, exact = TRUE)
-              cor_type <- paste0("cor_", g)
-              cor_pars <- get_cornames(rnames, cor_type, brackets = FALSE)
-              cor_samples <- matrix(
-                0, nrow = nrow(sd_samples), ncol = length(cor_pars)
-              )
-              for (k in seq_along(cor_pars)) {
-                if (cor_pars[k] %in% colnames(samples)) {
-                  cor_samples[, k] <- get_samples(
-                    samples, cor_pars[k], exact = TRUE
-                  )
-                }
-              }
-              cov_matrix <- get_cov_matrix(sd_samples, cor_samples)
-              # sample new levels from the normal distribution
-              # implied by the covariance matrix
-              indices <- ((j - 1) * nranef + 1):(j * nranef)
-              new_rsamples[[i]][, indices] <- t(apply(
-                cov_matrix, 1, rmulti_normal, 
-                n = 1, mu = rep(0, length(sd_pars))
-              ))
-            }
-          }
-          max_level <- max_level + length(new_indices)
-        } else if (sample_new_levels == "uncertainty") {
-          new_rsamples[[i]] <- matrix(nrow = nrow(rsamples), ncol = nranef)
-          for (k in seq_len(nranef)) {
-            # sample values for the new level
-            indices <- ((k - 1) * nlevels + 1):(k * nlevels)
-            new_rsamples[[i]][, k] <- apply(
-              rsamples[, indices, drop = FALSE], 1, sample, size = 1
-            )
-          }
-          max_level <- max_level + 1
-          gf[[i]][gf[[i]] > nlevels] <- max_level
-        }
-      } else { 
-        new_rsamples[[i]] <- matrix(nrow = nrow(rsamples), ncol = 0)
-      }
-    }
-    new_rsamples <- do_call(cbind, new_rsamples)
-    # we need row major instead of column major order
-    sort_levels <- ulapply(seq_len(nlevels),
-      function(l) seq(l, ncol(rsamples), nlevels)
+    # generate samples for new levels
+    args_new_rsamples <- nlist(
+      ranef = g_ranef, gf, used_levels, old_levels, 
+      rsamples, samples, sample_new_levels
     )
-    rsamples <- rsamples[, sort_levels, drop = FALSE]
-    # new samples are already in row major order
+    new_rsamples <- do_call(get_new_rsamples, args_new_rsamples)
+    max_level <- attr(new_rsamples, "max_level")
+    gf <- attr(new_rsamples, "gf")
     rsamples <- cbind(rsamples, new_rsamples)
+    # keep only those levels actually used in the current data
     levels <- unique(unlist(gf))
     rsamples <- subset_levels(rsamples, levels, nranef)
+    
+    # store samples and corresponding data in 'draws'
     # special group-level terms (mo, me, mi)
-    sub_ranef_sp <- subset2(sub_ranef, type = "sp")
-    if (nrow(sub_ranef_sp)) {
+    g_ranef_sp <- subset2(g_ranef, type = "sp")
+    if (nrow(g_ranef_sp)) {
       Z <- matrix(1, length(gf[[1]])) 
       draws[["Zsp"]][[g]] <- prepare_Z(Z, gf, max_level, weights)
-      for (co in sub_ranef_sp$coef) {
-        take <- which(sub_ranef$coef == co & sub_ranef$type == "sp")
+      for (co in g_ranef_sp$coef) {
+        take <- which(g_ranef$coef == co & g_ranef$type == "sp")
         take <- take + nranef * (seq_along(levels) - 1)
         draws[["rsp"]][[co]][[g]] <- rsamples[, take, drop = FALSE]
       }
     }
     # category specific group-level terms
-    sub_ranef_cs <- subset2(sub_ranef, type = "cs")
-    if (nrow(sub_ranef_cs)) {
+    g_ranef_cs <- subset2(g_ranef, type = "cs")
+    if (nrow(g_ranef_cs)) {
       # all categories share the same Z matrix
-      take <- grepl("\\[1\\]$", sub_ranef_cs$coef)
-      Znames <- paste0(
-        "Z_", sub_ranef_cs$id[take], usc(p), "_", sub_ranef_cs$cn[take]
-      )
+      take <- grepl("\\[1\\]$", g_ranef_cs$coef)
+      Znames <- paste0("Z_", g_ranef_cs$id[take], p, "_", g_ranef_cs$cn[take])
       Z <- do_call(cbind, sdata[Znames])
       draws[["Zcs"]][[g]] <- prepare_Z(Z, gf, max_level, weights)
       for (i in seq_len(sdata$ncat - 1)) {
         index <- paste0("\\[", i, "\\]$")
-        take <- which(grepl(index, sub_ranef$coef) & sub_ranef$type == "cs")
+        take <- which(grepl(index, g_ranef$coef) & g_ranef$type == "cs")
         take <- as.vector(outer(take, nranef * (seq_along(levels) - 1), "+"))
         draws[["rcs"]][[g]][[i]] <- rsamples[, take, drop = FALSE]
       }
     }
-    # basic group-level effects
-    sub_ranef_basic <- subset2(sub_ranef, type = c("", "mmc"))
-    if (nrow(sub_ranef_basic)) {
-      Znames <- paste0(
-        "Z_", sub_ranef_basic$id, usc(p), "_", sub_ranef_basic$cn
-      )
-      if (sub_ranef_basic$gtype[1] == "mm") {
-        ng <- length(sub_ranef_basic$gcall[[1]]$groups)
+    # basic group-level terms
+    g_ranef_basic <- subset2(g_ranef, type = c("", "mmc"))
+    if (nrow(g_ranef_basic)) {
+      Znames <- paste0("Z_", g_ranef_basic$id, p, "_", g_ranef_basic$cn)
+      if (g_ranef_basic$gtype[1] == "mm") {
+        ng <- length(g_ranef_basic$gcall[[1]]$groups)
         Z <- vector("list", ng)
         for (k in seq_len(ng)) {
           Z[[k]] <- do_call(cbind, sdata[paste0(Znames, "_", k)])
@@ -678,7 +597,7 @@ extract_draws_re <- function(bterms, samples, sdata, data, ranef, old_ranef,
         Z <- do_call(cbind, sdata[Znames])
       }
       draws[["Z"]][[g]] <- prepare_Z(Z, gf, max_level, weights)
-      take <- which(sub_ranef$type %in% c("", "mmc"))
+      take <- which(g_ranef$type %in% c("", "mmc"))
       take <- as.vector(outer(take, nranef * (seq_along(levels) - 1), "+"))
       rsamples <- rsamples[, take, drop = FALSE]
       draws[["r"]][[g]] <- rsamples
@@ -804,6 +723,7 @@ pseudo_draws_for_mixture <- function(draws, comp, sample_ids = NULL) {
 subset_levels <- function(x, levels, nranef) {
   # take relevant cols of a matrix of group-level terms
   # if only a subset of levels is provided (for newdata)
+  # requires x to be in row major order
   # Args:
   #   x: a matrix typically samples of r or Z design matrices
   #   levels: grouping factor levels to keep
@@ -814,13 +734,28 @@ subset_levels <- function(x, levels, nranef) {
   x[, take_levels, drop = FALSE]
 }
 
+column_to_row_major_order <- function(x, nranef) {
+  # transform x from column to row major order
+  # rows represent levels and columns represent effects
+  # Args:
+  #   x: a matrix of samples of group-level parameters
+  #   nranef: number of group-level effects
+  nlevels <- ncol(x) / nranef
+  sort_levels <- ulapply(seq_len(nlevels),
+    function(l) seq(l, ncol(x), by = nlevels)
+  )
+  x[, sort_levels, drop = FALSE]
+}
+
 prepare_Z <- function(Z, gf, max_level = NULL, weights = NULL) {
-  # prepare group-level design matrices for use in linear_predictor
+  # prepare group-level design matrices for use in 'predictor'
   # Args:
   #   Z: (list of) matrices to be prepared
   #   gf: (list of) vectors containing grouping factor values
   #   weights: optional (list of) weights of the same length as gf
   #   max_level: maximal level of gf
+  # Returns:
+  #   a sparse matrix representation of Z
   if (!is.list(Z)) {
     Z <- list(Z)
   }
@@ -868,6 +803,108 @@ expand_matrix <- function(A, x, max_level = max(x), weights = 1) {
     i = i, j = j, x = as.vector(t(A)),
     dims = c(nrow(A), ncol(A) * max_level)
   )
+}
+
+get_new_rsamples <- function(ranef, gf, rsamples, used_levels, old_levels,
+                             sample_new_levels, samples = NULL) {
+  # generate samples for new group levels
+  # Args:
+  #   ranef: 'ranef_frame' object of only a single grouping variable
+  #   gf: list of vectors of level indices in the current data
+  #   rsamples: matrix of group-level samples in row major order
+  #   used_levels: names of levels used in the current data
+  #   old_levels: names of levels used in the original data
+  #   sample_new_levels: specifies the way in which new samples are generated
+  #   samples: optional matrix of samples from all model parameters
+  # Returns:
+  #   a matrix of samples for new group levels
+  snl_options <- c("uncertainty", "gaussian", "old_levels")
+  sample_new_levels <- match.arg(sample_new_levels, snl_options)
+  g <- unique(ranef$group)
+  stopifnot(length(g) == 1L)
+  stopifnot(is.list(gf))
+  used_by_per_level <- attr(used_levels, "by")
+  old_by_per_level <- attr(old_levels, "by")
+  new_levels <- setdiff(used_levels, old_levels)
+  nranef <- nrow(ranef)
+  nlevels <- length(old_levels)
+  max_level <- nlevels
+  
+  out <- vector("list", length(gf))
+  for (i in seq_along(gf)) {
+    has_new_levels <- any(gf[[i]] > nlevels)
+    if (has_new_levels) {
+      if (sample_new_levels %in% c("old_levels", "gaussian")) {
+        new_indices <- sort(setdiff(gf[[i]], seq_len(nlevels)))
+        out[[i]] <- matrix(NA, nrow(rsamples), nranef * length(new_indices))
+        if (sample_new_levels == "old_levels") {
+          for (j in seq_along(new_indices)) {
+            # choose an existing person to take the parameters from
+            if (length(old_by_per_level)) {
+              new_by <- used_by_per_level[used_levels == new_levels[j]]
+              possible_levels <- old_levels[old_by_per_level == new_by]
+              possible_levels <- which(old_levels %in% possible_levels)
+              take_level <- sample(possible_levels, 1)
+            } else {
+              take_level <- sample(seq_len(nlevels), 1)
+            }
+            for (k in seq_len(nranef)) {
+              take <- (take_level - 1) * nranef + k
+              out[[i]][, (j - 1) * nranef + k] <- rsamples[, take]
+            }
+          }
+        } else if (sample_new_levels == "gaussian") {
+          if (any(!ranef$dist %in% "gaussian")) {
+            stop2("Option sample_new_levels = 'gaussian' is not ",
+                  "available for non-gaussian group-level effects.")
+          }
+          for (j in seq_along(new_indices)) {
+            # extract hyperparameters used to compute the covariance matrix
+            if (length(old_by_per_level)) {
+              new_by <- used_by_per_level[used_levels == new_levels[j]]
+              rnames <- as.vector(get_rnames(ranef, bylevels = new_by))
+            } else {
+              rnames <- get_rnames(ranef)
+            }
+            sd_pars <- paste0("sd_", g, "__", rnames)
+            sd_samples <- get_samples(samples, sd_pars, exact = TRUE)
+            cor_type <- paste0("cor_", g)
+            cor_pars <- get_cornames(rnames, cor_type, brackets = FALSE)
+            cor_samples <- matrix(0, nrow(sd_samples), length(cor_pars))
+            for (k in seq_along(cor_pars)) {
+              if (cor_pars[k] %in% colnames(samples)) {
+                cor_samples[, k] <- get_samples(
+                  samples, cor_pars[k], exact = TRUE
+                )
+              }
+            }
+            cov_matrix <- get_cov_matrix(sd_samples, cor_samples)
+            # sample new levels from the normal distribution
+            # implied by the covariance matrix
+            indices <- ((j - 1) * nranef + 1):(j * nranef)
+            out[[i]][, indices] <- t(apply(
+              cov_matrix, 1, rmulti_normal, 
+              n = 1, mu = rep(0, length(sd_pars))
+            ))
+          }
+        }
+        max_level <- max_level + length(new_indices)
+      } else if (sample_new_levels == "uncertainty") {
+        out[[i]] <- matrix(nrow = nrow(rsamples), ncol = nranef)
+        for (k in seq_len(nranef)) {
+          indices <- seq(k, nlevels * nranef, by = nranef)
+          tmp <- rsamples[, indices, drop = FALSE]
+          out[[i]][, k] <- apply(tmp, 1, sample, size = 1)
+        }
+        max_level <- max_level + 1
+        gf[[i]][gf[[i]] > nlevels] <- max_level
+      }
+    } else { 
+      out[[i]] <- matrix(nrow = nrow(rsamples), ncol = 0)
+    }
+  }
+  out <- do_call(cbind, out)
+  structure(out, gf = gf, max_level = max_level)
 }
 
 get_samples <- function(x, pars, ...) {