+++ R/family-lists.R
@@ -141,6 +141,29 @@
   )
 }
 
+.family_discrete_weibull <- function() {
+  list(
+    links = c(
+      "logit", "probit", "probit_approx",
+      "cloglog", "cauchit", "identity"
+    ),
+    dpars = c("mu", "shape"), type = "int", 
+    ybounds = c(0, Inf), closed = c(TRUE, NA),
+    ad = c("weights", "subset", "cens", "trunc"),
+    include = "fun_discrete_weibull.stan"
+  )
+}
+
+.family_com_poisson <- function() {
+  list(
+    links = c("log", "identity", "sqrt"),
+    dpars = c("mu", "shape"), type = "int", 
+    ybounds = c(0, Inf), closed = c(TRUE, NA),
+    ad = c("weights", "subset", "cens", "trunc"),
+    include = "fun_com_poisson.stan"
+  )
+}
+
 .family_gamma <- function() {
   list(
     links = c("log", "identity", "inverse"),